custom-elements/registering-custom-elements/unresolved-elements-interface-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>Unresolved element interface must be SVGElement, if the namespace is SVG Namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="When an unresolved element is created, it's element interface must be SVGElement if the namespace is SVG Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();

    var xfoo = doc.createElementNS('http://www.w3.org/2000/svg', 'x-foo');
    
    assert_class_string(xfoo, 'SVGElement', 'Element must be a SVG element');
    
}, 'Unresolved element interface with valid name for SVG namespace');

</script>
</body>
</html>
